prepare for PR. added comments.  cleaned up tests
