+++ jest.config.js
@@ -4,6 +4,6 @@ module.exports = {
    },
    testRegex: "^.+\\.spec\\.ts$",
    moduleFileExtensions: ["ts", "js"],
-   collectCoverage: false,
+   collectCoverage: true,
    testEnvironment: "node",
 };
