--- src/lib/ejs/skipped.ejs
@@ -1,28 +1,0 @@
-<hr>
- <h2>Skipped</h2>
-    <% if (!!skip) { %>
-        <%  Object.keys(skip).forEach(function(key) { %>      
-            <div class="wrap-collabsible">
-                <input id="<%= key %>skip" class="toggle" type="checkbox">
-                <label for="<%= key %>skip" class="lbl-toggle skip"><%= key %></label>
-                <div class="collapsible-content">
-                    <div class="content-inner">
-                    <table>
-                        <tr>
-                            <th>Scenario</th>
-                            <th>Required By</th>
-                        </tr>
-                        <% for (const feature of skip[key])  { -%>
-                            <% if (!!feature.scenarioName) { -%>
-                                <tr>
-                                    <td class="skip"><%= feature.scenarioName %></td>
-                                    <td class="skip"><%= feature.requiredBy %></td>
-                                </tr>
-                            <% } -%>
-                        <% } -%>             
-                    </table>
-                    </div>
-                </div>
-            </div>
-        <% }); -%>
-    <% } -%>
